Leetcode/src/leetcode/no094_BinaryTreeInorderTraversal/Solution.java
package leetcode.no094_BinaryTreeInorderTraversal;

import java.util.ArrayDeque;
import java.util.ArrayList;
import java.util.Deque;
import java.util.List;

import leetcode.TreeNode;

public class Solution {
	public List<Integer> inorderTraversal(TreeNode root) {
		List<Integer> inorder = new ArrayList<>();
		Deque<TreeNode> stack = new ArrayDeque<>();
		TreeNode node = root;
		while (!stack.isEmpty() || node != null) {
			if (node != null) {
				stack.push(node);
				node = node.left;
			} else {
				node = stack.pop();
				inorder.add(node.val);
				node = node.right;
			}
		}
		return inorder;
	}
}
